FIXME needed local copy of CudaKernelSources.h
make_float3 and similar CUDA vector operations are defined
in CudaKernelSources::VectorOps, however those are just
created on the fly by CMake at build time and not installed
by OpenMM.
Temporary solution was to copy a generated version of this
to this folder.
Is there a better way of doing this?